PR Requests: Move dependencies to correct directory
