+++ hpobench/benchmarks/ml_mmfb/nn_benchmark.py
@@ -6,7 +6,7 @@ import numpy as np
 from ConfigSpace.hyperparameters import Hyperparameter
 from sklearn.neural_network import MLPClassifier
 
-from hpobench.dependencies.ml.ml_benchmark_template import MLBenchmark
+from hpobench.dependencies.ml_mmfb.ml_benchmark_template import MLBenchmark
 
 
 class NNBenchmark(MLBenchmark):
